feat: add more content to homepage
