+++ .vscode/settings.json
@@ -1,5 +1,8 @@
 {
   "rust-analyzer.linkedProjects": ["./Cargo.toml"],
   "editor.formatOnSave": true,
-  "editor.defaultFormatter": "rust-lang.rust-analyzer"
+  "editor.defaultFormatter": "rust-lang.rust-analyzer",
+  "[html]": {
+    "editor.defaultFormatter": "esbenp.prettier-vscode"
+  }
 }
