+++ src/services/home_service.rs
@@ -1,3 +1,13 @@
-pub async fn home() -> &'static str {
-    "Welcome to Informatics FRS Helper API"
+use askama::Template;
+use axum::response::IntoResponse;
+
+use super::HtmlTemplate;
+
+#[derive(Template)]
+#[template(path = "home.html")]
+struct HomeTemplate;
+
+pub async fn home() -> impl IntoResponse {
+    let template = HomeTemplate {};
+    HtmlTemplate(template)
 }
