+++ src/services/mod.rs
@@ -1,4 +1,9 @@
-use axum::{http::StatusCode, Json};
+use askama::Template;
+use axum::{
+    http::StatusCode,
+    response::{Html, IntoResponse},
+    Json,
+};
 use serde::Serialize;
 use serde_json::{json, Value};
 pub mod class_service;
@@ -52,5 +57,23 @@ fn display_err(variant: ErrorViews) -> (StatusCode, String) {
             (StatusCode::NOT_FOUND, format!("{} tidak ditemukan", field))
         }
         ErrorViews::BadRequest(message) => (StatusCode::BAD_REQUEST, message),
+    }
+}
+
+pub struct HtmlTemplate<T>(T);
+
+impl<T> IntoResponse for HtmlTemplate<T>
+where
+    T: Template,
+{
+    fn into_response(self) -> axum::response::Response {
+        match self.0.render() {
+            Ok(html) => Html(html).into_response(),
+            Err(err) => (
+                StatusCode::INTERNAL_SERVER_ERROR,
+                format!("Failed to render template. Error: {err}"),
+            )
+                .into_response(),
+        }
     }
 }
