+++ templates/home.html
@@ -1,0 +1,46 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Home - API Informatics FRS Helper</title>
+  </head>
+  <body>
+    <h1>API Informatics FRS Helper</h1>
+    <p>
+      Ini adalah API Public yang dapat digunakan untuk mendapatkan informasi
+      tambahan terkait FRS yang mungkin tidak ditemukan di website ini.
+      Informasi yang bisa didapatkan terkait <strong>daftar kelas</strong>,
+      <strong>dosen</strong>, dan <strong>mata kuliah</strong>.
+    </p>
+    <p>Dibuat oleh: <a href="https://github.com/albugowy15">albugowy15</a></p>
+
+    <h2>Dokumentasi</h2>
+    <p>
+      Klik link dibawah untuk melihat dokumentasi sekaligus mencoba api
+      menggunakan swagger.
+    </p>
+    <a href="/swagger">Swagger</a>
+
+    <h2>Resources</h2>
+    <ul>
+      <li><a href="https://tc-frs-helper.live">FRS Helper Website</a></li>
+      <li>
+        <a
+          href="https://github.com/albugowy15/api-informatics-frs-helper"
+          target="_blank"
+          >Github</a
+        >
+      </li>
+      <li>
+        <a
+          href="https://documenter.getpostman.com/view/30505077/2s9YsJBCJo"
+          target="_blank"
+          >Postman Documentation</a
+        >
+      </li>
+    </ul>
+
+    <small>Copyright 2024 - Mohamad Kholid Bughowi</small>
+  </body>
+</html>
