+++ templates/swagger.html
@@ -1,21 +1,27 @@
 <!DOCTYPE html>
 <html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Dokumentasi Swagger API Informatics FRS Helper</title>
-    <link rel="stylesheet" href="https://unpkg.com/swagger-ui-dist@4.5.0/swagger-ui.css" />
-</head>
-<body>
+    <link
+      rel="stylesheet"
+      href="https://unpkg.com/swagger-ui-dist@4.5.0/swagger-ui.css"
+    />
+  </head>
+  <body>
     <div id="swagger-ui"></div>
-    <script src="https://unpkg.com/swagger-ui-dist@4.5.0/swagger-ui-bundle.js" crossorigin></script>
+    <script
+      src="https://unpkg.com/swagger-ui-dist@4.5.0/swagger-ui-bundle.js"
+      crossorigin
+    ></script>
     <script>
-    window.onload = () => {
+      window.onload = () => {
         window.ui = SwaggerUIBundle({
-        url: '{{ open_api_json_url }}',
-        dom_id: '#swagger-ui',
+          url: "/assets/openapi.json",
+          dom_id: "#swagger-ui",
         });
-    };
+      };
     </script>
-</body>
+  </body>
 </html>
